Updated style prop issues references
The old referenced issue #7070 is fairly outdated, multiple styling issues reported back then are now working correctly (e.g. borderYXRadius styling props).
The issue #30809 is an updated bug report for issues affecting TextInput in more recent versions of React Native (tested and confirmed on 0.63.2). It also has reference screenshots for developers to quickly understand the issues.
This part of the documentation should be updated when the issues presented in #30809 are resolved.